+++ src/ts/Components/Particles/Particles.ts
@@ -1,4 +1,6 @@
-export function ColorContrastClass(color?:string, contrast?:string, type?: 'txt' | 'bg') {
+import { ColorValues, ContrastValues, LayoutMargin, LayoutType, Spacing } from "../../types"
+
+export function ColorContrastClass(color?:ColorValues, contrast?:ContrastValues, type?: 'txt' | 'bg') {
 return [`${color}-${type}`, {
     'cont-hst': contrast === 'highest',
     'cont-h': contrast === 'high',
@@ -8,7 +10,7 @@ return [`${color}-${type}`, {
   }]
 }
 
-export function LayoutClass(type?:string, margin?:string) {
+export function LayoutClass(type?:LayoutType, margin?:LayoutMargin) {
   return [{
     'a-Layout-full': type === 'full',
     'a-Layout-center': type === 'center',
@@ -23,4 +25,22 @@ export function LayoutClass(type?:string, margin?:string) {
     'a-Layout-margin-single': margin === 'single',
     'a-Layout-margin-none': margin === 'none'
   }]
+}
+
+export function BoxMarginClass(margin?:Spacing) {
+  return {
+    'a-Box-m0': margin === 'none',
+    'a-Box-m1': margin === 'single',
+    'a-Box-m2': margin === 'double',
+    'a-Box-m3': margin === 'triple',
+  }
+}
+
+export function ListGapClass(gap?:Spacing) {
+  return {
+    'a-List-g0': gap === 'none',
+    'a-List-g1': gap === 'single',
+    'a-List-g2': gap === 'double',
+    'a-List-g3': gap === 'triple',
+  }
 }